+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.0] - 2022-10-18
+### Added
+- coordinates and sensor layout XML files for the Geodesic Head Web
+
+### Fixed
+- writing of categories.xml when `channelStatus` attr is `None`
+
 ## [0.7.4] - 2022-06-22
 ### Fixed
 - lxml future warning
@@ -83,7 +90,8 @@ In `Reader.get_physical_samples_from_epoch()`:
 - Parse key elements in categories.xml files with `mffpy.xml_files.Categories` class.
 - Incorporate `cached_property` dependency into `mffpy` library.
 
-[Unreleased]: https://github.com/bel-public/mffpy/compare/v0.7.4...HEAD
+[Unreleased]: https://github.com/bel-public/mffpy/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/bel-public/mffpy/compare/v0.7.4...v0.8.0
 [0.7.4]: https://github.com/bel-public/mffpy/compare/v0.7.3...v0.7.4
 [0.7.3]: https://github.com/bel-public/mffpy/compare/v0.7.2...v0.7.3
 [0.7.2]: https://github.com/bel-public/mffpy/compare/v0.7.1...v0.7.2